requirements.txt
pynamodb_attributes==0.3.1
Flask==1.1.2
pynamodb==5.1.0
gunicorn==20.1.0
pyyaml